TW-34970 ensure only one thread executes addBuilds()
The exception occurs during comparison of an old state of the queue with the new one to calculate builds reorderings. Since we ensured old state contains only builds in the new state, the only way it can happen is when another thread calls addBuilds() and alters the state of the object (it is called on server startup without any synchronization). This method relied on synchronization in BuildQueue, now it has its own sync.